+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New optional `file` option for the `[logging]` section in the webservice
 configuration. Can be used to additionally output all log messages to the given
 file.
+- `Graph:ensure_loaded_parallel` returns the actually loaded components that did
+exist.
 
 ### Fixed